+++ packages/multichain-account-service/src/tests/messenger.ts
@@ -45,6 +45,8 @@ export function getMultichainAccountServiceMessenger(
       'KeyringController:getState',
       'KeyringController:getKeyringsByType',
       'KeyringController:addNewKeyring',
+      'NetworkController:findNetworkClientIdByChainId',
+      'NetworkController:getNetworkClientById',
     ],
   });
 }